(#35) Validation on classes
Durations should be float values (half hour, etc).

We check if classes are scheduled in the future as part of validation.

Need a mechanism to check overlap for instructors / rooms.

+++ src/main/java/edu/njit/cs631/fitness/data/entity/Clazz.java
@@ -104,11 +104,11 @@ public class Clazz {
 	}
     
 	@Column(name="DURATION")
-	private Integer duration;
-	public Integer getDuration() {
+	private Double duration;
+	public Double getDuration() {
 		return duration;
 	}
-	public void setDuration(Integer duration) {
+	public void setDuration(Double duration) {
 		this.duration = duration;
 	}
